refactoring core logic into sampling module
