+++ fiftyone/brain/__init__.py
@@ -92,3 +92,56 @@ def compute_uniqueness(samples, uniqueness_field="uniqueness", roi_field=None):
     fbu.compute_uniqueness(
         samples, uniqueness_field=uniqueness_field, roi_field=roi_field,
     )
+
+
+def sample_best_video_frames(
+    video_path,
+    out_frames_dir,
+    target_num_frames=None,
+    target_accel=None,
+    target_fps=None,
+    size=None,
+    max_size=None,
+):
+    """Adaptively samples the best frames from the input video.
+
+    The "best" video frames at a given sampling density are defined as the
+    frames with highest image quality that are most representative of the
+    visual content in the video.
+
+    Provide one of ``target_num_frames``, ``target_accel``, or ``target_fps``
+    to perform the sampling.
+
+    Args:
+        video_path: the path to the video to process
+        out_frames_dir: a directory to write the sampled frames
+        target_num_frames (None): the target number of frames to sample
+        target_accel (None): a desired target acceleration factor to apply when
+            sampling frames. For example, a target acceleration of 2x would
+            correspond to sampling every other frame, on average
+        target_fps (None): a desired target sampling rate, which must be less
+            than the frame rate of the input video
+        size (None): a desired output ``(width, height)`` for the sampled
+            frames. Dimensions can be -1, in which case the input aspect ratio
+            is preserved. By default, the input frame size is maintained
+        max_size (None): a maximum ``(width, height)`` allowed for the sampled
+            frames. Frames are resized as necessary to meet this limit, and
+            ``size`` is decreased (aspect-preserving) if necessary to satisfy
+            this constraint. Dimensions can be -1, in which case no limit is
+            applied to them. By default, no maximum frame size is imposed
+
+    Returns:
+        a dict mapping frame numbers to the paths to the sampled frames in
+        ``out_frames_dir``
+    """
+    import fiftyone.brain.internal.core.sampling as fbs
+
+    return fbs.sample_best_video_frames(
+        video_path,
+        out_frames_dir,
+        target_num_frames=target_num_frames,
+        target_accel=target_accel,
+        target_fps=target_fps,
+        size=size,
+        max_size=max_size,
+    )
